Moving db files into config directory
