+++ gunicorn_config_api.py
@@ -1,5 +1,5 @@
 from api import app
-from db_config import initialize_database
+from src.config.db_config import initialize_database
 
 
 def on_starting(server):
